Optimize handlers and ui_overlay

+++ src/utils/ui_overlay.cpp
@@ -13,12 +13,12 @@ namespace ui_overlay {
 		screen.pen = Pen(255, 255, 255, 100);
 		screen.rectangle(Rect(1, screen.bounds.h - 10, 20, 9));
 		screen.pen = Pen(255, 255, 255, 200);
-		std::string fms = std::to_string((int) (1 / ((ms_end - ms_start) / 1000)));
+		std::string fms = std::to_string(static_cast<int>(1 / ((ms_end - ms_start) / 1000)));
 		screen.text(fms, minimal_font, Rect(3, screen.bounds.h - 9, 10, 16));
 
 		int block_size = 4;
-		for (uint32_t i = 0; i < (ms_end - ms_start); i++) {
-			screen.pen = Pen((int) i * 5, 255 - ((int) i * 5), 0);
+		for (int i = 0; i < static_cast<int>(ms_end - ms_start); i++) {
+			screen.pen = Pen(i * 5, 255 - (i * 5), 0);
 			screen.rectangle(Rect(i * (block_size + 1) + 1 + 21, screen.bounds.h - block_size - 1, block_size, block_size));
 		}
 	}
